refactor(frontend): dividir CardDetailModal em componentes menores
- Extrair CardModalHeader, CardModalFooter, CardFormFields e ImageModal
- Refatorar CardDetailModal para usar componentes modulares
- Melhorar manutenibilidade e reutilização de código
- Simplificar estrutura do modal principal
